refactor: de-clutter main simulator file

+++ simulator/smt/simulator.py
@@ -1,133 +1,55 @@
 #!/usr/bin/env python
 
+import os, sys, shutil
 import multiprocessing
-import os, sys, re, shutil
-from taskFactory import Task
+from parserFactory import Aggregator
 from tasksetGenerator import Generator
 from virtualGangFactory import VirtualGangCreator
 
 SEED = 5
-PRISTINE = True
+PRISTINE = False
 NUM_OF_CORES = 8
 EDGE_PROBABILITY = 25
 MAX_TASKS_PER_PERIOD = 8
-NUM_OF_TEST_TASKSETS = 1000
+NUM_OF_TEST_TASKSETS = 10
 UTILIZATIONS = range(1, NUM_OF_CORES)
 RESULT_FILE = 's%d_vgangs.txt' % (SEED)
-PARALLELISM = multiprocessing.cpu_count() * 2
+PARALLELISM = multiprocessing.cpu_count()
 
 def main():
-    if PRISTINE:
-        processes = {}
+    if PRISTINE: parallel_create_virtual_taskset()
 
-        # Remove generated directory for pristine execution
-        generated_dir = os.getcwd() + '/generated'
-        if os.path.exists(generated_dir): shutil.rmtree(generated_dir)
-
-        print
-        for r in range(0, NUM_OF_TEST_TASKSETS, PARALLELISM):
-            for tsIdx in range(r, min(r + PARALLELISM, NUM_OF_TEST_TASKSETS)):
-                processes[tsIdx] = multiprocessing.Process(target = \
-                        form_virtual_gangs, args = (tsIdx,))
-
-                processes[tsIdx].start()
-
-            for tsIdx in range(r, min(r + PARALLELISM, NUM_OF_TEST_TASKSETS)):
-                processes[tsIdx].join()
-
-        print '\n'
-
-    tasksets = aggregate_tasksets()
+    # Aggregate results by parsing the file-system logs and re-create real and
+    # virtual tasksets for further processing
+    aggregator = Aggregator()
+    tasksets = aggregator.run()
     dbg_dump_vgang_info(tasksets)
 
     return
 
-def aggregate_tasksets():
-    '''
-      Traverse the directory containing generated tasksets and their respective
-      virtual-gangs data for further processing.
-    '''
-    assert os.path.exists('generated'), ("Directory of generated data does "
-        "not exists in the current folder.")
+def parallel_create_virtual_taskset():
+    processes = {}
 
-    tasksets = {}
-    generated_taskset_dirs = os.listdir('generated')
+    # Remove generated directory for pristine execution
+    generated_dir = os.getcwd() + '/generated'
+    if os.path.exists(generated_dir): shutil.rmtree(generated_dir)
 
-    for ts in generated_taskset_dirs:
-        tsIdx, util, period = parse_taskset_dir(ts)
+    print
+    for r in range(0, NUM_OF_TEST_TASKSETS, PARALLELISM):
+        for tsIdx in range(r, min(r + PARALLELISM, NUM_OF_TEST_TASKSETS)):
+            processes[tsIdx] = multiprocessing.Process(target = \
+                    virtual_gang_generator_thread_entry, args = (tsIdx,))
 
-        if not tasksets.has_key(tsIdx):
-            tasksets[tsIdx] = {}
+            processes[tsIdx].start()
 
-        if not tasksets[tsIdx].has_key(util):
-            tasksets[tsIdx][util] = {}
+        for tsIdx in range(r, min(r + PARALLELISM, NUM_OF_TEST_TASKSETS)):
+            processes[tsIdx].join()
 
-        if not tasksets[tsIdx][util].has_key(period):
-            tasksets[tsIdx][util][period] = {}
+    print '\n'
 
-        tasksets[tsIdx][util][period]['Real'] = parse_taskset(ts, 'real')
-        tasksets[tsIdx][util][period]['Virtual'] = parse_taskset(ts, 'virtual')
+    return
 
-    return tasksets
-
-def parse_taskset(taskset_dir, nature):
-    name_prefix = 'candidate_' if nature == 'real' else 'virtual_task'
-    taskset_file = 'generated/%s/%sset.txt' % (taskset_dir, name_prefix)
-
-    assert os.path.exists(taskset_file), ("Taskset file <%s> does "
-            "not exists in the generated directory." % (taskset_file))
-
-    pattern = '^Task:\s+([\d]+) \| C=\s*([\d.]+) P=\s*([\d]+) h=\s*([\d]+)' + \
-        ' r=\s*([\d]+) u=\s*([\d.]+) ([^\s]+)$'
-
-    taskset = []
-    with open(taskset_file, 'r') as fdi:
-        lines = fdi.readlines()
-
-        for l in lines:
-            m = re.match(pattern, l)
-
-            if not m:
-                continue
-
-            tid = int(m.group(1))
-            C = float(m.group(2))
-            p = int(m.group(3))
-            h = int(m.group(4))
-            r = int(m.group(5))
-            u = float(m.group(6))
-
-            e = []
-            members = ''
-            if nature == 'real': e = parse_edges(m.group(7))
-            if nature == 'virtual': members = m.group(7)[2:]
-
-            taskset.append(Task(tid, C, p, h, r, e, members))
-
-    return taskset
-
-def parse_edges(edge_string):
-    if edge_string == 'None':
-        return []
-
-    pattern = r'->t([\d]+)'
-    edge_list = [int(e) for e in re.findall(pattern, edge_string)]
-
-    return edge_list
-
-def parse_taskset_dir(ts_dir):
-    pattern = r'ts([\d]+)_u([\d]+)_p([\d]+)'
-    m = re.match(pattern, ts_dir)
-
-    assert m, ("Taskset directory <%s> does not match expected pattern.")
-
-    tsIdx = int(m.group(1))
-    util = int(m.group(2))
-    period = int(m.group(3))
-
-    return tsIdx, util, period
-
-def form_virtual_gangs(tsIdx):
+def virtual_gang_generator_thread_entry(tsIdx):
     # Generate taskset and then create SMT script
     taskFactory = Generator(NUM_OF_CORES, UTILIZATIONS, EDGE_PROBABILITY,
             tsIdx + 1, MAX_TASKS_PER_PERIOD)
@@ -165,17 +87,6 @@ def form_virtual_gangs(tsIdx):
 
     return
 
-def dbg_dump_vgang_info(tasksets):
-    rem_result_file()
-
-    for tsIdx in tasksets:
-        with open(RESULT_FILE, 'a') as fdo:
-            print_taskset(fdo, tasksets[tsIdx], tsIdx)
-
-            fdo.write('\n')
-
-    return
-
 def print_taskset(fdo, taskset, tsIdx):
     fdo.write("Taskset # %d\n" % (tsIdx))
 
@@ -195,18 +106,23 @@ def print_taskset(fdo, taskset, tsIdx):
 
     return
 
+def dbg_dump_vgang_info(tasksets):
+    if os.path.exists(RESULT_FILE): os.remove(RESULT_FILE)
+
+    for tsIdx in tasksets:
+        with open(RESULT_FILE, 'a') as fdo:
+            print_taskset(fdo, tasksets[tsIdx], tsIdx)
+
+            fdo.write('\n')
+
+    return
+
 def print_progress(cur_taskset, max_tasksets, cur_util, max_util, period):
     print '[PROGRESS] Processing Taskset: %4d / %-4d | Utilization: %2d '   \
             '/ %-2d | Period: %4d\r' % (cur_taskset, max_tasksets, cur_util, \
                     max_util, period),
 
     sys.stdout.flush()
-
-    return
-
-def rem_result_file():
-    if os.path.exists(RESULT_FILE):
-        os.remove(RESULT_FILE)
 
     return
